feat: add recipients data to calling remote mute event
