+++ src/script/calling/CallingRepository.ts
@@ -71,11 +71,12 @@ import {Config} from '../Config';
 import {isGroupMLSConversation, isMLSConversation, MLSConversation} from '../conversation/ConversationSelectors';
 import {ConversationState} from '../conversation/ConversationState';
 import {ConversationVerificationState} from '../conversation/ConversationVerificationState';
-import {CallingEvent, EventBuilder} from '../conversation/EventBuilder';
+import {EventBuilder} from '../conversation/EventBuilder';
 import {CONSENT_TYPE, MessageRepository, MessageSendingOptions} from '../conversation/MessageRepository';
 import {Conversation} from '../entity/Conversation';
 import type {User} from '../entity/User';
 import {NoAudioInputError} from '../error/NoAudioInputError';
+import {CallingEvent} from '../event/CallingEvent';
 import {EventRepository} from '../event/EventRepository';
 import {EventSource} from '../event/EventSource';
 import type {MediaDevicesHandler} from '../media/MediaDevicesHandler';
